fix: production-readiness bug fixes and split rules.go monolith
Bug fixes:
- engine: fix Windows infinite loop in AnalyzeWorkflow's repo-root walk. The
  loop terminated on the Unix root "/" only, so filepath.Dir("C:\\") never
  matched and any single-workflow scan outside a git repo hung indefinitely.
- parser: accept templated `continue-on-error: ${{ }}` (the field was typed
  bool, so a common real-world construct aborted the entire scan). Also skip a
  single unparseable workflow with a warning instead of failing the whole scan.
- organization: recalculate the org summary after AI verification (was a no-op
  self-assignment) and implement generateTopFindings (was a stub returning []).
- rules/context: gate ARTIPACKED HIGH severity on an artifact upload that can
  include the .git directory; the context analyzer no longer raises severity
  based on untrusted input (irrelevant to credential-persistence risk). This
  removes HIGH false positives on essentially every actions/checkout.
- report: deduplicate findings before computing the summary so the headline
  issue count matches the JSON/SARIF output (was 19 vs 16 on real repos).
- Dockerfile: build with golang:1.25 to match go.mod, and copy default
  policies from the builder stage (the previous check ran in the final stage
  where no source exists, so it was always a no-op).

Refactor:
- split the 4,784-line pkg/rules/rules.go into cohesive category files
  (secrets, supply_chain_checks, triggers, malicious_patterns, permissions,
  runners, debug, conditions). Behaviour-neutral; verified by existing tests.
